pyproject.toml
[build-system]
# Keeping setuptools strictly below 70.0 as requested
requires = ["setuptools>=61.0,<70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pysoftk"
version = "1.0.0"
description = "PySoftK: Python Soft-Matter Kings College London"
readme = "README.txt"
authors = [
  {name = "A. Santana-Bonilla"},
  {name = "R. Lopez-Castro"},
  {name = "R. Ziolek"},
  {name = "C. Lorenz", email = "k2031560@kcl.ac.uk"}
]
requires-python = ">=3.7"
dependencies = [
    "pytest >= 2.7.2",
    "numpy",
    "Cython",  # Kept as a runtime dep since it was in your original list
    "networkx >= 2.7; python_version >= '3.8'",
    "networkx < 2.7; python_version < '3.8'",
    "rdkit",
    "wheel",
    "tqdm >= 4.64.0",
    "pathos >= 0.2.8",
    "pyscf < 2.5.0",
    "pyscf-semiempirical @ git+https://github.com/pyscf/semiempirical",
    "pyberny",
    "openbabel-wheel",
    "numba",
    "umap-learn",
    "hdbscan",
    "pandas",
    "pyarrow",
    "MDAnalysis",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["test*"]